requirements.txt
streamlit==1.37.0
psycopg2-binary==2.9.9
pandas==2.2.2
sentence-transformers==3.0.1
transformers==4.44.2
safetensors==0.4.4
torch==2.4.1
requests==2.32.3
pydantic==2.8.2
python-dotenv==1.0.1
numpy>=2.0,<3

